+++ packages/plugin-commons/web/src/Icons/index.ts
@@ -1,8 +1,5 @@
 // @create-index
 
-export { default as AlignCenterIcon } from './AlignCenterIcon';
-export { default as BackArrowIcon } from './BackArrowIcon';
-export { default as HeightIcon } from './HeightIcon';
 export { default as PluginSettingsIcon } from './PluginSettingsIcon';
 export { default as SizeBestFitIcon } from './SizeBestFitIcon';
 export { default as SizeFullWidthIcon } from './SizeFullWidthIcon';
@@ -16,8 +13,6 @@ export { default as SizeSmallRightIcon } from './SizeSmallRightIcon';
 export { default as SpoilerButtonIcon } from './SpoilerButtonIcon';
 export { default as TrashIcon } from './TrashIcon';
 export { default as WidthIcon } from './WidthIcon';
-export { default as EditIcon } from './EditIcon';
 export { default as RTLIcon } from './RTLIcon';
 export { default as LTRIcon } from './LTRIcon';
 export { default as LoaderIcon } from './LoaderIcon';
-export { default as HashtagIcon } from './HashtagIcon';