+++ packages/plugin-html/web/src/icons/index.ts
@@ -1,3 +1,2 @@
-export { default as EditIcon } from './EditIcon';
 export { default as InsertPluginIcon } from './InsertPluginIcon';
 export { default as AdsenseIcon } from './Adsense';