+++ packages/toolbars/web/src/icons/index.ts
@@ -1,1 +1,0 @@
-export { default as DropdownArrowIcon } from './DropdownArrowIcon';