+++ packages/ui-components/web/src/Icons/index.ts
@@ -1,11 +1,11 @@
 // @create-index
 
+export { default as DropdownArrowIcon } from './DropdownArrowIcon';
 export { default as InfoIcon } from './InfoIcon';
-
-// plugin-commons
-// export { default as AlignCenterIcon } from './AlignCenterIcon';
-// export { default as BackArrowIcon } from './BackArrowIcon';
-// export { default as HeightIcon } from './HeightIcon';
+export { default as AddColorIcon } from './AddColorIcon';
+export { default as AlignCenterIcon } from './AlignCenterIcon';
+export { default as HeightIcon } from './HeightIcon';
+export { default as ErrorIcon } from './ErrorIcon';
 // export { default as PluginSettingsIcon } from './PluginSettingsIcon';
 // export { default as SizeBestFitIcon } from './SizeBestFitIcon';
 // export { default as SizeFullWidthIcon } from './SizeFullWidthIcon';
@@ -20,13 +20,12 @@ export { default as InfoIcon } from './InfoIcon';
 // export { default as TrashIcon } from './TrashIcon';
 // export { default as WidthIcon } from './WidthIcon';
 export { default as CloseIcon } from './CloseIcon';
-// export { default as EditIcon } from './EditIcon';
+export { default as EditIcon } from './EditIcon';
 // export { default as RTLIcon } from './RTLIcon';
 // export { default as LTRIcon } from './LTRIcon';
 // export { default as LoaderIcon } from './LoaderIcon';
-// export { default as HashtagIcon } from './HashtagIcon';
+export { default as HashtagIcon } from './HashtagIcon';
 export { default as MoreIcon } from './MoreIcon';
 export { default as ClearIcon } from './ClearIcon';
 export { default as SearchIcon } from './SearchIcon';
-export { default as DropdownArrowIcon } from './DropdownArrowIcon';
 export { default as CheckIcon } from './CheckIcon';